aegis-cli/pyproject.toml
[project]
name = "aegis-cli"
version = "0.1.0"
description = "CLI tool for Agentic Ops"
authors = [
    {name = "User", email = "user@example.com"},
]
dependencies = [
    "typer[all]",
    "rich",
    "httpx",
    "pyyaml"
]
requires-python = ">=3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
aegis = "src.main:app"
